+++ src/hooks/useFileContextMenu.ts
@@ -1,7 +1,6 @@
 import createContextMenu from "../components/ContextMenu/createContextMenu.ts";
-import {WebviewWindow} from "@tauri-apps/api/window";
 import useSelection from "./useSelection.ts";
-import {compareType, openFile} from "../utils";
+import {openFile} from "../utils";
 
 const useFileContextMenu = () => {
   const selection = useSelection();
@@ -14,26 +13,7 @@ const useFileContextMenu = () => {
           shortcut: "F5",
           handler() {
             const file = Array.from(selection.items.value)[0];
-
-            if (!compareType(file.fileSuffix, "image")) {
-              openFile(file.fullPath)
-            } else {
-              const fileWindow = WebviewWindow.getByLabel("file")
-              if (fileWindow) {
-                fileWindow.emit("update_file", {id: file.id})
-                fileWindow.setFocus()
-              } else {
-                const webview = new WebviewWindow("file", {
-                  url: `/file/${file.id}`,
-                  decorations: false,
-                  title: "查看文件",
-                  visible: false
-                });
-                webview.once("page_loaded", () => {
-                  webview.show()
-                })
-              }
-            }
+            openFile(file)
           },
         }
       ]
